Conflicts:
	director-common/src/main/java/com/intel/director/common/Constants.java
	director-common/src/main/java/com/intel/director/common/MountImage.java
	director-trust-policy/src/main/java/com/intel/mtwilson/director/trust/policy/CreateTrustPolicy.java
	director-trust-policy/src/main/java/com/intel/mtwilson/director/trust/policy/DirectoryAndFileUtil.java
	features/director-async-image-processor/src/main/java/com/intel/director/async/ImageActionTaskFactory.java
	features/director-async-image-processor/src/main/java/com/intel/director/async/task/CreateTarTask.java
	features/director-async-image-processor/src/main/java/com/intel/director/async/task/ExecuteActionsTask.java
	features/director-async-image-processor/src/main/java/com/intel/director/async/task/UploadTarTask.java
	features/director-async-image-processor/src/main/java/com/intel/director/async/task/UploadTask.java
	features/director-async-image-processor/src/main/java/com/intel/director/quartz/ImageActionPoller.java
	features/director-async-image-processor/src/main/java/com/intel/director/quartz/ImageActionScheduler.java
	features/director-glance/pom.xml
	features/director-glance/src/main/java/com/intel/director/images/rs/GlanceRsClient.java
	features/director-html5/src/main/html5/mtwilson-core-html5/navbar/settings.html
	features/director-html5/src/main/html5/mtwilson-core-html5/navbar/td_images.html
	features/director-html5/src/main/html5/public/bm_images_grid_page.html
	features/director-html5/src/main/html5/public/confirmation_bare_metal_live.html
	features/director-html5/src/main/html5/public/create_policy_select_bm_directories.html
	features/director-html5/src/main/html5/public/create_policy_select_directories.html
	features/director-html5/src/main/html5/public/create_policy_wizard.html
	features/director-html5/src/main/html5/public/css/font-awesome.min.css
	features/director-html5/src/main/html5/public/file_upload_window.html
	features/director-html5/src/main/html5/public/images/ajax-loader.gif
	features/director-html5/src/main/html5/public/js/bm_images_grid_page.js
	features/director-html5/src/main/html5/public/js/create_policy.js
	features/director-html5/src/main/html5/public/js/create_policy_bm_image.js
	features/director-html5/src/main/html5/public/js/create_policy_for_hosts.js
	features/director-html5/src/main/html5/public/js/create_policy_select_directories.js
	features/director-html5/src/main/html5/public/js/create_policy_select_directories_for_hosts.js
	features/director-html5/src/main/html5/public/js/create_policy_wizard_docker.js
	features/director-html5/src/main/html5/public/js/edit_policy.js
	features/director-html5/src/main/html5/public/js/edit_policy_bm_live.js
	features/director-html5/src/main/html5/public/js/edit_policy_wizard_docker.js
	features/director-html5/src/main/html5/public/js/trust_policy.js
	features/director-html5/src/main/html5/public/js/upload_to_bare_metal.js
	features/director-html5/src/main/html5/public/js/upload_to_image_store.js
	features/director-html5/src/main/html5/public/js/upload_to_image_store_bm.js
	features/director-html5/src/main/html5/public/js/vm_images_grid_page.js
	features/director-html5/src/main/html5/public/js/vm_images_page.js
	features/director-html5/src/main/html5/public/setting_create_ssh.html
	features/director-html5/src/main/html5/public/settings_upload_image_to_director.html
	features/director-html5/src/main/html5/public/trust_policy_main.html
	features/director-html5/src/main/html5/public/upload_imagestore_bm_image_direct.html
	features/director-html5/src/main/html5/public/upload_imagestore_direct.html
	features/director-html5/src/main/html5/public/upload_imagestore_docker.html
	features/director-html5/src/main/html5/public/upload_imagestore_docker_direct.html
	features/director-html5/src/main/html5/public/vm_images_grid_page.html
	features/director-html5/src/main/html5/public/vm_images_page.html
	features/director-images-api/src/main/java/com/intel/director/api/CreateTrustPolicyMetaDataRequest.java
	features/director-images-api/src/main/java/com/intel/director/api/CreateTrustPolicyResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/GenericResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/ImageActionRequest.java
	features/director-images-api/src/main/java/com/intel/director/api/ImageActionResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/ImageInfoResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/ImportPolicyTemplateResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/ListImageLaunchPolicyResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/SshSettingResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/TrustPolicyResponse.java
	features/director-images-api/src/main/java/com/intel/director/api/UpdateTrustPolicyRequest.java
	features/director-images-ws/src/main/java/com/intel/director/images/ImageActions.java
	features/director-images-ws/src/main/java/com/intel/director/images/Images.java
	features/director-images-ws/src/main/java/com/intel/director/images/Setting.java
	features/director-images-ws/src/main/java/com/intel/director/images/TrustPolicies.java
	features/director-images-ws/src/main/java/com/intel/director/images/TrustPolicyDrafts.java
	features/director-images-ws/src/main/java/com/intel/director/service/ImageService.java
	features/director-images-ws/src/main/java/com/intel/director/service/impl/ImageActionImpl.java
	features/director-images-ws/src/main/java/com/intel/director/service/impl/ImageServiceImpl.java
	features/director-images-ws/src/main/java/com/intel/director/service/impl/LookupServiceImpl.java
	features/director-images-ws/src/main/java/com/intel/director/service/impl/SSHManager.java
	features/director-images-ws/src/main/java/com/intel/director/util/TdaasUtil.java
	features/director-persistence/src/main/java/com/intel/mtwilson/director/dao/ImageDao.java
	features/director-persistence/src/main/java/com/intel/mtwilson/director/data/MwImage.java
	features/director-persistence/src/main/java/com/intel/mtwilson/director/dbservice/DbServiceImpl.java
	features/director-persistence/src/test/java/com/intel/mtwilson/director/db/DbServiceTest.java
	packages/director-zip/pom.xml
	packages/director-zip/src/main/application/bin/director.sh
	packages/director/src/main/resources/psql.sql
	packages/director/src/main/resources/setup.sh

+++ features/director-images-ws/src/main/java/com/intel/director/images/ImageActions.java
@@ -15,13 +15,20 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
+import org.apache.commons.lang.StringUtils;
+
+import com.intel.dcsg.cpg.validation.RegexPatterns;
+import com.intel.dcsg.cpg.validation.ValidationUtil;
+import com.intel.director.api.GenericResponse;
 import com.intel.director.api.ImageActionObject;
 import com.intel.director.api.ImageActionRequest;
 import com.intel.director.api.ImageActionResponse;
-import com.intel.director.common.Constants;
+import com.intel.director.api.ui.ImageInfo;
 import com.intel.director.images.exception.DirectorException;
 import com.intel.director.service.ImageActionService;
+import com.intel.director.service.ImageService;
 import com.intel.director.service.impl.ImageActionImpl;
+import com.intel.director.service.impl.ImageServiceImpl;
 import com.intel.mtwilson.launcher.ws.ext.V2;
 
 /**
@@ -35,81 +42,81 @@ import com.intel.mtwilson.launcher.ws.ext.V2;
 public class ImageActions {
 
 	ImageActionService actionService = new ImageActionImpl();
-
+	ImageService imageService = new ImageServiceImpl();
 	private static final org.slf4j.Logger log = org.slf4j.LoggerFactory
 			.getLogger(ImageActions.class);
 
 	/**
-	 * This method will fetch an image-action which has actionId on which
-	 * actions are performed, list of actions to be performed, etc.
-	 * In case no image action find with given id return HTTP 404 Not Found
-	 * 
-	 * action_count and action_completed are provided for convenience. These two
-	 * attributes are guaranteed to have the right data corresponding to the
-	 * actions collection.
-	 * 
-	 * The tasks in "action" attribute are processed sequentially. The current
-	 * task holds the name of the task in the "action" array that is currently
-	 * being processed.
-	 * 
+	  This method will fetch an image-action based on actionId on which
+	  actions are performed, list of actions to be performed, etc.
+	  if actionId do not exist it give 404 Not Found.
+	  
+	  action_count and action_completed are provided for convenience. These two
+	  attributes are guaranteed to have the right data corresponding to the
+	  actions collection. action size - is the data uploaded, action_size_max 
+	  - total size of the image uploaded
+	  
+	  The tasks in "action" attribute are processed sequentially. The current
+	  task holds the name of the task in the "action" array that is currently
+	  being processed.
+	  
 	 * @mtwContentTypeReturned JSON
 	 * @mtwMethodType GET
 	 * @mtwSampleRestCall <pre>
-	 * https://server.com:8443/v1/image-actions
-	 * Input: action_id = CF0A8FA3-F73E-41E9-8421-112FB22BB057
-	 * Output: {
-	 * "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
-	 * "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
-	 * "action_count": 2,
-	 * "action_completed": 2,
-	 * "action_size": 66570,
-	 * "action_size_max": 66570,
-	 * "action": [ { "status": "Complete","task_name": "Create Tar"},
-	 *  "status": "Complete", "storename": "Glance", "task_name": "Upload Tar" }],
-	 * "current_task_status": "Complete",
-	 * "current_task_name": "Upload Tar" }
-	 * 
-	 *  {
-	 * "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
-	 * "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
-	 * "action_count": 2,
-	 * "action_completed": 1,
-	 * "action_size": 66570,
-	 * "action_size_max": 66570,
-	 * "action": [ { "status": "Complete","task_name": "Create Tar"},
-	 *  "status": "In Progress", "storename": "Glance", "task_name": "Upload Tar" }],
-	 * "current_task_status": "Complete",
-	 * "current_task_name": "Create Tar" }
-	 * 
-	 * 
-	 * In case of error creating tar : 
-	 * {
-	 * "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
-	 * "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
-	 * "action_count": 2,
-	 * "action_completed": 0,
-	 * "action_size": 66570,
-	 * "action_size_max": 66570,
-	 * "action": [ { "status": "Error","task_name": "Create Tar","error":"Error Creating tar"},
-	 * "status": "Incomplete", "storename": "Glance", "task_name": "Upload Tar" }],
-	 * "current_task_status": "Error : Error Creating tar ",
-	 * "current_task_name": "Create Tar" }
-	 * 
-	 * In case of error uploading tar : 
-	 * {
-	 * "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
-	 * "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
-	 * "action_count": 2,
-	 * "action_completed": 1,
-	 * "action_size": 66570,
-	 * "action_size_max": 66570,
-	 * "action": [ { "status": "Complete","task_name": "Create Tar"},
-	 * "status": "Error", "storename": "Glance", "task_name": "Upload Tar","error":"Error Uploading tar"}],
-	 * "current_task_status": "Error : Error Uploading tar ",
-	 * "current_task_name": "Upload Tar" }
-	 * 
-	 * In case no image action find with given id return HTTP 404 Not Found
-	 * </pre>
+	  https://{IP/HOST_NAME}/v1/image-actions/<action_id>
+	  Input: PathParam : action_id = CF0A8FA3-F73E-41E9-8421-112FB22BB057
+	  Output: {
+	  "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
+	  "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
+	  "action_count": 2,
+	  "action_completed": 2,
+	  "action_size": 66570,
+	  "action_size_max": 66570,
+	  "action": [ { "status": "Complete","task_name": "Create Tar"},
+	   "status": "Complete", "storename": "Glance", "task_name": "Upload Tar" }],
+	  "current_task_status": "Complete",
+	  "current_task_name": "Upload Tar" }
+	  
+	   {
+	  "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
+	  "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
+	  "action_count": 2,
+	  "action_completed": 1,
+	  "action_size": 66570,
+	  "action_size_max": 66570,
+	  "action": [ { "status": "Complete","task_name": "Create Tar"},
+	   "status": "In Progress", "storename": "Glance", "task_name": "Upload Tar" }],
+	  "current_task_status": "Complete",
+	  "current_task_name": "Create Tar" }
+	  
+	  
+	  In case of error creating tar : 
+	  {
+	  "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
+	  "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
+	  "action_count": 2,
+	  "action_completed": 0,
+	  "action_size": 66570,
+	  "action_size_max": 66570,
+	  "action": [ { "status": "Error","task_name": "Create Tar","error":"Error Creating tar"},
+	  "status": "Incomplete", "storename": "Glance", "task_name": "Upload Tar" }],
+	  "current_task_status": "Error : Error Creating tar ",
+	  "current_task_name": "Create Tar" }
+	  
+	  In case of error uploading tar : 
+	  {
+	  "id": "CF0A8FA3-F73E-41E9-8421-112FB22BB057",
+	  "image_id": "08EB37D7-2678-495D-B485-59233EB51996",
+	  "action_count": 2,
+	  "action_completed": 1,
+	  "action_size": 66570,
+	  "action_size_max": 66570,
+	  "action": [ { "status": "Complete","task_name": "Create Tar"},
+	  "status": "Error", "storename": "Glance", "task_name": "Upload Tar","error":"Error Uploading tar"}],
+	  "current_task_status": "Error : Error Uploading tar ",
+	  "current_task_name": "Upload Tar" }
+	  
+	  </pre>
 	 * 
 	 * @param actionId
 	 * @return ImageActionResponse
@@ -120,25 +127,38 @@ public class ImageActions {
 	@GET
 	public Response fetchImageAction(@PathParam("actionId") String actionId) {
 		try {
-			return Response.ok(actionService.fetchImageAction(actionId))
-					.build();
-		} catch (DirectorException e) {
-			return Response.status(404)
+			GenericResponse genericResponse= new GenericResponse();
+			if(!ValidationUtil.isValidWithRegex(actionId,RegexPatterns.UUID)){
+				genericResponse.error = "Action Id is empty or not in uuid format";
+				return Response.status(Response.Status.BAD_REQUEST)
+						.entity(genericResponse).build();
+			}
+			if(actionService.fetchImageAction(actionId)!=null){
+				return Response.ok(actionService.fetchImageAction(actionId))
+						.build();
+					
+			}else{
+				return  Response.status(404)
 					.entity("No image action found for id: " + actionId)
 					.build();
+			}
+			
+		} catch (DirectorException e) {
+			return Response.status(500).build();
 		}
 	}
 
 	/**
 	 * This method will create an image-action. Data required by this method is
-	 * action_id and list of task and other parameter associated with it(Ex.
-	 * store_name in case of Upload Tar task). Output contains actionId(id),image_id, actions, etc.
+	 * image_id and list of task and other parameter associated with it(Ex.
+	 * store_name in case of Upload Tar task). Output contains actionId(id),image_id, actions etc.
+	 * 
 	 * 
 	 * 
 	 * @mtwContentTypeReturned JSON
 	 * @mtwMethodType POST
 	 * @mtwSampleRestCall <pre>
-	 * https://server.com:8443/v1/image-actions
+	 * https://{IP/HOST_NAME}/v1/image-actions
 	 * Input: { "image_id":"08EB37D7-2678-495D-B485-59233EB51996",
 	 * "actions":[ {"task_name":"Create Tar"},
 	 * {"task_name":"Upload Tar","storename":"Glance"}]
@@ -164,14 +184,10 @@ public class ImageActions {
 	 * 	"current_task_name": "Create Tar" 
 	 * }
 	 * 
-	 * In case of error:
-	 * {
-	 * 		"error" : "Error in create ImageAction"
-	 * }
 	 * 
 	 * </pre>
 	 * 
-	 * @param ImageActionRequest
+	 * @param imageActionRequest 
 	 * @return ImageActionResponse containing action_id.
 	 */
 
@@ -179,19 +195,39 @@ public class ImageActions {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	@POST
-	public ImageActionResponse createImageAction(
+	public Response createImageAction(
 			ImageActionRequest imageActionRequest) {
 		ImageActionResponse imageActionResponse = new ImageActionResponse();
 		ImageActionObject imageActionObject;
+		String error=imageActionRequest.vaidate();
+		if(!StringUtils.isBlank(error)){
+			imageActionResponse.error=error;
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(imageActionResponse).build();
+		}
+		
+		ImageInfo fetchImageById = null;
+		try {
+			fetchImageById = imageService
+					.fetchImageById(imageActionRequest.image_id);
+		} catch (DirectorException e1) {
+			log.error("unable to fetch image");
+		}
+
+		if (fetchImageById == null) {
+			imageActionResponse.error = "Image does not exist for id: "
+					+ imageActionRequest.image_id;
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(imageActionResponse).build();
+		}
+		
 		try {
 			imageActionObject = actionService
 					.createImageAction(imageActionRequest);
 		} catch (DirectorException e) {
 			log.error("Error in createImageAction", e);
-			imageActionResponse.setStatus(Constants.ERROR);
-			imageActionResponse.setDetails(e.getMessage());
 			imageActionResponse.setError(e.getMessage());
-			return imageActionResponse;
+			return Response.ok(imageActionResponse).build();
 		}
 		imageActionResponse.setId(imageActionObject.getId());
 		imageActionResponse.setImage_id(imageActionObject.getImage_id());
@@ -204,7 +240,7 @@ public class ImageActions {
 				.getCurrent_task_name());
 		imageActionResponse.setCurrent_task_status(imageActionObject
 				.getCurrent_task_status());
-		return imageActionResponse;
+		return Response.ok(imageActionResponse).build();
 	}
 
 	/**
@@ -214,33 +250,48 @@ public class ImageActions {
 	 * @mtwContentTypeReturned JSON
 	 * @mtwMethodType DELETE
 	 * @mtwSampleRestCall <pre>
-	 * https://server.com:8443/v1/image-actions
+	 * https://{IP/HOST_NAME}/v1/image-actions
 	 * Input: PathParam =  actionId : CF0A8FA3-F73E-41E9-8421-112FB22BB057
-	 * Output: {"deleted": true}
+	 * Output: {"deleted":true,"action_count":0,"action_completed":0,"action_size":0,"action_size_max":0}
 	 * 
-	 * In case of error:
-	 * {“deleted”: false, “error”: “No image action with id :: CF0A8FA3-F73E-41E9-8421-112FB22BB057” }
 	 * </pre>
 	 * 
-	 * @param imageActionRequest
 	 * @return Status of delete operation
 	 */
 	@Path("image-actions/{actionId: [0-9a-zA-Z_-]+}")
 	@Produces(MediaType.APPLICATION_JSON)
 	@DELETE
-	public ImageActionResponse deleteImageAction(
+	public Response deleteImageAction(
 			@PathParam("actionId") String actionId) {
 		ImageActionResponse imageActionResponse = new ImageActionResponse();
 		imageActionResponse.setDeleted(true);
+	///	GenericResponse genericResponse= new GenericResponse();
+		if(!ValidationUtil.isValidWithRegex(actionId,RegexPatterns.UUID)){
+			imageActionResponse.error = "Action Id is empty or not in uuid format";
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(imageActionResponse).build();
+		}
+		ImageActionObject fetchImageAction = null;
+		try {
+			fetchImageAction = actionService.fetchImageAction(actionId);
+		} catch (DirectorException e1) {
+			log.error("unable to fetch image action", e1);
+		}
+
+		if(fetchImageAction == null){			
+			imageActionResponse.error = "Image action does not exist for id "+actionId;
+			return Response.status(Response.Status.BAD_REQUEST)
+					.entity(imageActionResponse).build();
+		}
 		try {
 			actionService.deleteImageAction(actionId);
-		} catch (DirectorException e) {
+		} catch (Exception e) {
 			log.error("Error in deleteImageAction", e);
+			imageActionResponse.setError("Error in deleteImageAction");
 			imageActionResponse.setDeleted(false);
-			imageActionResponse.setDetails(e.getMessage());
-			return imageActionResponse;
+			return Response.ok(imageActionResponse).build();
 		}
-		return imageActionResponse;
+		return Response.ok(imageActionResponse).build();
 	}
 
 }
